test: keep merged coverage files within limits
