+++ pkg/services/models/transports/cli/bootstrap_invoke_test.go
@@ -14,6 +14,7 @@ import (
 	modelinference "github.com/portpowered/infinite-you/pkg/services/models"
 	operatorconfig "github.com/portpowered/infinite-you/pkg/services/operator_settings"
 	"github.com/portpowered/infinite-you/pkg/services/work"
+	"github.com/portpowered/infinite-you/pkg/transports/cli/resolvedinput"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 	apisurface "github.com/portpowered/infinite-you/pkg/transports/mapping"
 	"go.uber.org/zap"
@@ -458,5 +459,44 @@ func TestInvoke_AudioBootstrapCopiesStreamFile(t *testing.T) {
 	}
 	if string(got) != "RIFF....WAVE" {
 		t.Fatalf("output = %q, want streamed audio bytes", string(got))
+	}
+}
+
+func TestModelsCLIMissingCacheMessageKeepsThePullHint(t *testing.T) {
+	t.Parallel()
+
+	if got := modelCacheNotFoundMessage(modelinference.ErrModelCacheNotFound); got != "model cache is not installed; run you models pull <model> first" {
+		t.Fatalf("modelCacheNotFoundMessage() = %q, want actionable pull hint", got)
+	}
+}
+
+func TestReadModelsInvokeInputsClearsOnlyManifestOperationDefaultForNamedInputs(t *testing.T) {
+	t.Parallel()
+
+	inputs, err := resolvedinput.Resolve(
+		[]resolvedinput.Definition{
+			{ID: modelsInvokeNameInputID, Kind: resolvedinput.ValueKindString, Precedence: []resolvedinput.Source{resolvedinput.SourcePositionalArgument}},
+			{ID: modelsInvokeOperationID, Kind: resolvedinput.ValueKindString, Precedence: []resolvedinput.Source{resolvedinput.SourceManifestDefault, resolvedinput.SourceCLIFlag}},
+			{ID: modelsInvokeTextID, Kind: resolvedinput.ValueKindString, Precedence: []resolvedinput.Source{resolvedinput.SourceManifestDefault}},
+			{ID: modelsInvokeInputID, Kind: resolvedinput.ValueKindStringArray, Precedence: []resolvedinput.Source{resolvedinput.SourceCLIFlag}},
+			{ID: modelsInvokeOutputID, Kind: resolvedinput.ValueKindStringArray, Precedence: []resolvedinput.Source{resolvedinput.SourceManifestDefault}},
+		},
+		[]resolvedinput.Candidate{
+			{InputID: modelsInvokeNameInputID, Source: resolvedinput.SourcePositionalArgument, Value: resolvedinput.StringValue("embed")},
+			{InputID: modelsInvokeOperationID, Source: resolvedinput.SourceManifestDefault, Value: resolvedinput.StringValue("TTS")},
+			{InputID: modelsInvokeTextID, Source: resolvedinput.SourceManifestDefault, Value: resolvedinput.StringValue("")},
+			{InputID: modelsInvokeInputID, Source: resolvedinput.SourceCLIFlag, Value: resolvedinput.StringArrayValue([]string{"text=hello"})},
+			{InputID: modelsInvokeOutputID, Source: resolvedinput.SourceManifestDefault, Value: resolvedinput.StringArrayValue(nil)},
+		},
+	)
+	if err != nil {
+		t.Fatalf("resolve invoke inputs: %v", err)
+	}
+	got, err := readModelsInvokeInputs(inputs)
+	if err != nil {
+		t.Fatalf("readModelsInvokeInputs() error = %v", err)
+	}
+	if got.modelName != "embed" || got.operation != "" || len(got.inputMappings) != 1 || got.inputMappings[0] != "text=hello" {
+		t.Fatalf("readModelsInvokeInputs() = %#v, want cleared default operation and named input", got)
 	}
 }
